pyproject.toml
[project]
name = "ascendo-conf-agent"
version = "0.1.0"
description = "Conference scraper + prospect fit triage agent"
requires-python = ">=3.10"
dependencies = [
  "langgraph>=0.2.0",
  "pydantic>=2.6.0",
  "playwright>=1.45.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.2.0",
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
  "python-dotenv>=1.0.0",
  "google-genai>=0.3.0",
  "rich>=13.7.0",
  "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
